refactor(actions): Split functions into different files
